Accommodating NY Eastern District
The NY Eastern District publishes its own locally-produced RSS feed that contains no additional information, but is formatted slightly differently and has a non-standard URL. This URL will be manually added in the PACER Tracker admin for now. This update ensures that manually-inserted information will not be overridden by the loadcourts management command.